refactored impl to improve readability
